fix: consolidate vite config files to restore production build (#49)
Merge vite.config.mjs settings (base, serveDataPlugin, publicDir) into
vite.config.js and delete vite.config.mjs. Vite v6 resolves .js before
.mjs, causing the production-critical settings to be ignored since PR #46.
